requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
frozenlist==1.8.0
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
mysql-connector-python==9.6.0
mysqlclient==2.2.8
numpy==2.2.6
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
propcache==0.4.1
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
schedule==1.2.2
six==1.17.0
twilio==9.10.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
yarl==1.23.0
